% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{a_latent_resid_set}
\alias{a_latent_resid_set}
\title{a_latent_resid_set}
\usage{
a_latent_resid_set(vars, varname)
}
\arguments{
\item{vars}{Vector of variable names (column names in the target data.frame).}

\item{varname}{Used to name the latent residuals.}
}
\value{
A list with the model text in \code{model_text}, latent residual
  names in \code{varname}, and \code{type} set to 'latent' (not used yet).
}
\description{
Creates latent residual terms for each variable in \code{vars}, using
\code{varname} to name them.
}
