% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{a_random_intercept}
\alias{a_random_intercept}
\title{a_random_intercept}
\usage{
a_random_intercept(vars, varname)
}
\arguments{
\item{vars}{Vector of variable names (column names in the target data.frame).}

\item{varname}{Used to name the latent random intercept.}
}
\value{
A list with the model text in \code{model_text}, random intercept
  name in \code{varname}, and \code{type} set to 'latent' (not used yet).
}
\description{
Create syntax for a single random intercept term based on a vector of
variable names and a variable name for the random intercept term. Used
internally by \code{\link{random_intercepts}}.
}
