% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{clpm}
\alias{clpm}
\title{clpm}
\usage{
clpm(lat_resid_vars, constrain = TRUE)
}
\arguments{
\item{lat_resid_vars}{A list of vectors of variable names of the latent
residual variables. Each element should be a vector of variable names that
corresponds to the same construct measured repeatedly over waves.}

\item{constrain}{Constrain the path coefficiecients to be the same from
wave-to-wave?}
}
\value{
Model text specifying the CLPM portion of the model.
}
\description{
Create model text specifying the cross-lag panel model portion of the
RI-CLPM, including all autoregressive and cross-lag effects. Set
\code{constrain} to \code{TRUE} to constrain these to be equal across all
waves.
}
