% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{contemp_covars}
\alias{contemp_covars}
\title{contemp_covars}
\usage{
contemp_covars(lat_resid_vars, constrain = TRUE)
}
\arguments{
\item{lat_resid_vars}{A list of vectors of latent residual variable names.}

\item{constrain}{Constrain the covariances to be the same across waves?}
}
\value{
Model text specifying the contemporaneous covariances.
}
\description{
Create model text to specify contemporaneous covariances between latent
residuals.
}
