% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{latent_resids}
\alias{latent_resids}
\title{latent_resids}
\usage{
latent_resids(var_groups)
}
\arguments{
\item{var_groups}{A list of vectors of variable names. Each element of the
list will be used to generate model text for a set of latent residuals
based on the variables in that element.}
}
\value{
A matrix with columns for each top-level element of \code{var_groups}
  and rows for each element of the list returned by
  \code{\link{a_latent_resid_set}}.
}
\description{
Creates latent residuals model text for all sets of variables passed via
\code{var_groups}.
}
\examples{
some_vars <- list(x = c('x1', 'x2'), y = c('y1', 'y2'))
latent_resids(some_vars)
}
