% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{lavriclpm}
\alias{lavriclpm}
\title{lavriclpm}
\usage{
lavriclpm(riclpmModel, data, ...)
}
\arguments{
\item{riclpmModel}{The lavaan syntax for the RI-CLPM model}

\item{data}{The data}

\item{...}{Other parameters passed to \code{\link[lavaan]{lavaan}}}
}
\value{
a fitted model
}
\description{
This passes through a model and data to the lavaan::lavaan function with
proper constraints.
}
