% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{make_riclpm_lavaan}
\alias{make_riclpm_lavaan}
\title{make_riclpm_lavaan}
\usage{
make_riclpm_lavaan(xname = "x", yname = "y", waveints,
  constrain_all = F, alpha = "", delta = "", beta = "",
  gamma = "", uu = "", vv = "", uv = "", kk = "", oo = "",
  ko = "", pp = "", qq = "", pq = "")
}
\arguments{
\item{xname}{name of x variable}

\item{yname}{name of y variable}

\item{waveints}{integers specifying wave labels}

\item{constrain_all}{set all constraints via labels}

\item{alpha}{set label or value for alpha constraint}

\item{delta}{set label or value for delta constraint}

\item{beta}{set label or value for beta constraint}

\item{gamma}{set label or value for gamma constraint}

\item{uu}{set label or value for uu constraint}

\item{vv}{set label or value for vv constraint}

\item{uv}{set label or value for uv constraint}

\item{kk}{set label or value for kk constraint}

\item{oo}{set label or value for oo constraint}

\item{ko}{set label or value for ko constraint}

\item{pp}{set label or value for pp constraint}

\item{qq}{set label or value for qq constraint}

\item{pq}{set label or value for pq constraint}
}
\value{
A character string that is syntax for an RI-CLPM model as specified
}
\description{
make_riclpm_lavaan
}
