% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{manifest_ints}
\alias{manifest_ints}
\title{manifest_ints}
\usage{
manifest_ints(terms, constrain = "free")
}
\arguments{
\item{terms}{Vector of variable names to create intercept specifications for.}

\item{constraint}{Whether to constrain the intercepts to a certain value. \code{'free'} is the default. Provide the specific value for the contraint.}
}
\value{
Model text for specifying intercepts
}
\description{
Create model text that specifies intercepts for each term.
}
