% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{random_intercepts}
\alias{random_intercepts}
\title{random_intercepts}
\usage{
random_intercepts(var_groups)
}
\arguments{
\item{var_groups}{A list of vectors of variable names. Each element of the
list will be used to generate model text for a latent random intercept.}
}
\value{
A matrix with columns for each top-level element of \code{var_groups}
  and rows for each element of the list returned by
  \code{\link{a_random_intercept}}.
}
\description{
Creates random intercept model text for all sets of variables passed via
\code{var_groups}.
}
