% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{residual_variances}
\alias{residual_variances}
\title{residual_variances}
\usage{
residual_variances(var_groups, constrain = TRUE)
}
\arguments{
\item{var_groups}{A list of vectors of variable names. Each element of the
list will be used to generate model text for residual variances, possibly
with equality constraints across waves.}
}
\description{
Creat model text specifying residual variances, possibly with constraints to
eqaulity across waves.
}
