% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{riclpm_text}
\alias{riclpm_text}
\title{riclpm_text}
\usage{
riclpm_text(var_groups, constrain_over_waves = TRUE,
  constrain_ints = "free")
}
\arguments{
\item{var_groups}{A list of vectors of variable names.}

\item{constrain_over_waves}{Constrain regression coefficients, covariances, and residuals to be the same from wave to wave? Will not constrain variances and covariances of wave-1 latent residuals.}

\item{constrain_ints}{Constrain intercepts of manifest variables? Default is to free them. At the moment, passing other values is nonsense.}
}
\value{
Model text to be passed to a lavaan function.
}
\description{
Given a list of sets of variables sampled repeatedly over equal time intervals, create a Random Intercept Cross Lagged Panel Model (RI-CLPM) specification for lavaan.
}
