% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riclpmr.R
\docType{package}
\name{riclpmr}
\alias{riclpmr}
\alias{riclpmr-package}
\title{riclpmr: A package for creating RI-CLPM syntax for lavaan}
\description{
This riclpmr package is a set of tools for creating lavaan-style model
specifications of Random Intercept Cross Lag Panel Models for arbitrarily
large numbers of constructs and measurement occations.
}
\details{
It also contains a helpful function for passing this model specification to the appropriate \code{lavaan} or \code{blavaan} function.
}
