% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{simple_vars}
\alias{simple_vars}
\title{simple_vars}
\usage{
simple_vars(terms1, terms2, constrain = NULL)
}
\arguments{
\item{terms1}{A vector of variable names.}

\item{terms2}{A vector of variable names.}

\item{constrain}{Set all variances to equality. If not \code{FALSE}, provide
the name for the equality constraint.}
}
\value{
Model text specifying covariances between terms.
}
\description{
simple_vars
}
