% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{simriclpm}
\alias{simriclpm}
\title{simriclpm}
\usage{
simriclpm(riclpmModel, ...)
}
\arguments{
\item{riclpmModel}{Model text from \code{\link{riclpm_text}}.}

\item{...}{Additional parameters to \code{\link[lavaan]{simulateData}}.}
}
\description{
simriclpm
}
