% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syntax.R
\name{varcovars}
\alias{varcovars}
\title{varcovars}
\usage{
varcovars(terms)
}
\arguments{
\item{terms}{Vector of variable names to create variance-covariance
specification from.}
}
\value{
Model text specifying covariance between terms in \code{terms}.
}
\description{
Creates model text that specifies variances for each variable in
\code{terms}, as well as the full covariance matrix between those terms.
}
